# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "swagger documentation.",
        "title": "Best Lviv api docs",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/allocations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delivery"
                ],
                "summary": "List allocations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by request",
                        "name": "request_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to return",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/allocations/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delivery"
                ],
                "summary": "Approve a planned allocation (DISPATCHER)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Allocation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpresponse.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/entity.Allocation"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/allocations/{id}/dispatch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delivery"
                ],
                "summary": "Mark an allocation as dispatched (WORKER at source warehouse)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Allocation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpresponse.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/entity.Allocation"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/allocations/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delivery"
                ],
                "summary": "Reject an allocation (DISPATCHER)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Allocation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httprequest.RejectAllocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/v1/audit-log": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get audit log (ADMIN only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by actor",
                        "name": "actor_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by entity type",
                        "name": "entity_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to return",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/stock/nearest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delivery"
                ],
                "summary": "Find nearest warehouses with surplus stock for a resource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Destination customer ID",
                        "name": "point_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Required quantity (optional)",
                        "name": "quantity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Allocation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dispatched_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "request_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                },
                "source_warehouse_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/entity.AllocationStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.AllocationStatus": {
            "type": "string",
            "enum": [
                "planned",
                "approved",
                "in_transit",
                "delivered",
                "cancelled"
            ],
            "x-enum-varnames": [
                "AllocationStatusPlanned",
                "AllocationStatusApproved",
                "AllocationStatusInTransit",
                "AllocationStatusDelivered",
                "AllocationStatusCancelled"
            ]
        },
...
        "httprequest.RejectAllocationRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {
                    "$ref": "#/definitions/httpresponse.ResponseMetadata"
                }
            }
        },
        "httpresponse.ResponseMetadata": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token, \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}